Initial project setup
Minimal skeleton of the project to make linters happy. To prevent
shooting to our legs, we use mypy with dynamic typing disallowed
(i.e. no Any in type signatures). stubs is a directory with
annotations that are not yet present in mypy (annotations in stubs
are incomplete and they cover only needs of this project).