make coveralls work
Add a custom coveralls script to work around the problems with
coveralls and github actions.

Add a `coveralls26` testenv so that we can use coveralls with
py26.  This uses coveralls 1.x and coverage 4.x.

When running the github action, only use the tox environment corresponding
to the current system version of python installed by the github
action.  For example, if github action has installed the python v3.6,
we should only test py36-tox30, we should not test py27, py37, etc.
other python environments.

Get rid of references to Travis

Fix bug with iniconfig used with tox 3.20 and later.