improve collection testing
Get rid of `black` from collection testing.  It is too difficult
to make the generated code pass `black` testing.

Do not report `WARNINGS` from yamllint tests run against the
collection converted code.  It is too confusing for developers
to look at the tox log and determine which are "real" problems
they should be concerned about.

Instead of trying to reuse the .yamllint.yml etc. lint suppression
files from each repo, just create a single suppressions file that
is a superset of all of the suppressions from all of the repos.
Add environment variables like `RUN_YAMLLINT_CONFIG_FILE` so that
the collection test can override the lint check with the
suppressions used by collections.

When running `tox` in the collection directory, set the `--workdir`
to the original `repo/.tox/` directory so that ansible-doc and
ansible-lint won't try to analyze all of the files in the `.tox/`
directory.

Add additional `deps` required for ansible-doc and ansible-lint
tests.