support for ansible-test
Instead of using the environment variable
`LSR_ROLE2COLL_RUN_ANSIBLE_TESTS` to specify whether or not
to run `ansible-test`, create a separate testenv for it.  The
`ansible-test` env depends on `collection` to create the
collection.
Add the environment variables `LSR_ANSIBLE_TEST_DEBUG`,
`LSR_ANSIBLE_TEST_TESTS`, and `LSR_ANSIBLE_DOC_DEBUG`.
It was quite tricky to get `ansible-test` to work in a `tox`
environment . . .

Added a hack for developers - if you want to regenerate the
result.ini used by the merge_ini test, then set
`UPDATE_RESULT_INI=true tox`.  BE SURE TO CHECK THE RESULT
THAT IT IS CORRECT.