collections check for plugin fqcn
add `ansible-plugin-scan` to `-e collection` - the reason is that
in the collection format, plugins other than modules must use
FQCN
This also required the ability to set the `changedir` for the
tests that `-e collection` invokes using `LSR_RUN_TEST_DIR`
Use jinja 2.7 if python is 3.9 or lower to better test for jinja
constructs that are not supported on all platforms.
For ease of testing, allow using local collection and plugin scan
scripts by using `LSR_ROLE2COLL_DIR` and `LSR_REPORT_MODULES_PLUGINS`.

add support for python 3.10