add support for scraping centos download html page
Instead of periodically having to scan the centos images download
page: https://cloud.centos.org/centos/9-stream/x86_64/images/
to see if a new image has been posted, this will scrape the screen
HTML looking for a new image.  This requires a new `centoshtml`
image link type.  This works sort of like the `compose` image type
which looks for a new image in a compose.