add container testenv
This adds the ability to run a test using a container as a managed
host.  This depends on having `podman` installed on your local
system.  It also requires `jq`.

The interface was chosen to be as similar to `runqemu` as possible.
You run it like this:
`tox -e container-ansible-core-2.12 -- --image-name centos-8 tests/tests_default.yml`
This depends on https://github.com/linux-system-roles/linux-system-roles.github.io/pull/76
which adds containers for the image names, so that e.g. `--image-name centos-8`
is mapped to `quay.io/centos/centos:stream8`.
The script will build a local test image using the specified container as
a base image.  The test image will have a dnf/yum update, additional
repos configured, and any other setup steps performed.  This is useful if
you plan to run multiple tests and do not want to go through the setup
every time.  The test image will be automatically recreated if it is
too old.  You can force it to be recreated by using the
`--erase-old-snapshot` flag which works similar to `runqemu`.