add support for use-yum-cache, remove-cloud-init, use-snapshot, setup_yml, wait_on_qemu (#67)
`--use-yum-cache` - Cache files will be created which will be mounted
into the VM at `/var/cache/dnf` and `/var/cache/yum`, and
`/var/lib/dnf` and `/var/lib/yum`, which will cache packages and metadata
used by the test.  This should help speed up subsequent runs of the same
image - packages should be installed from the cache rather than the network.
This also requires that we use `yum makecache` or `dnf makecache`
during the setup phase.
`--remove-cloud-init` - Removes cloud-init and any dependencies it
pulls in from the VM before running the tests.  This is helpful to
determine missing package dependencies in the roles.
`--use-snapshot` will do the following:
* create a snapshot of the base image - the base image will become
  the backing image for the snapshot
* run the setup.yml against the snapshot to write to it, saving any
  initial setup - packages, repos, other setup - so that the setup
  step does not have to be performed upon subsequent runs using
  the image
* recreate the snapshot if it is older than 1 day or older than the
  backing file
`--wait-on-qemu` - Use the `LOCK_ON_FILE` feature of standard-inventory-qcow2
so that we can be sure qemu is shutdown before calling it again.
`--setup-yml` - setup tasks in addition to the setup from `config`
`--profile-task-limit` - the number of tasks to report when using `--profile`

This also adds support for running on EL7.

This also adds support for running with test-harness and other
environments other than tox.

This also adds EPEL 7 and 8, and disables them.  Roles that need
EPEL can just enable them.