add support for post-snap-sleep-time, shutdown of VM
Add support for `--post-snap-sleep-time` to control the time that
the script sleeps after shutting down the VM snapshot.  The
default value is `1` second.

Ensure the VM is shutdown cleanly by using `shutdown now` to shutdown
the VM rather than just killing qemu.

This also relies on
https://pagure.io/fork/rmeggins/standard-test-roles/c/f7831911cad699b2c19db869cc9d774c60f78c0e?branch=linux-system-roles
to help ensure the filesystems are synced.

sync on the host as well

make a copy of image env - do not use directly

write snap artifacts to artifacts.snap