support collection-requirements.yml (#78)
There is no way for a role to automatically require collections for
module/plugin dependencies.  Use the non-standard file
`meta/collection-requirements.yml` for this.  Note that
`meta/requirements.yml` is supposed to be a list of roles only, in
the format described at https://docs.ansible.com/ansible/latest/galaxy/user_guide.html#installing-multiple-roles-from-multiple-files
If we find a `meta/requirements.yml` file, and it is not in `list`
format, assume it is an old one that has not yet been renamed to
`meta/collection-requirements.yml`, use it, and warn the user.