collection - save, force, and restore local collection
If you are developing a system role that has a dependency on
fedora.linux_system_roles, and you want to test the collection,
you have to first convert your role to a collection under .tox/
as the fedora.linux_system_roles collection.  But you may also
need to install that collection from Galaxy in order to use other
roles.  In that case, save a copy of the local converted collection,
then do a force install of fedora.linux_system_roles, then restore
the local copy of the converted collection.