pytest: Create xml coverage report for codecov
Create the XML coverage report since this is the one that is needed for
codecov.

Note: codecov uses some heuristics to find the coverage file and it does
not seem to be possible to specify which file to use. The
get_coverage.yml playbook is also wrongly identified as a coverage
report which can be confusing if the actual report is not identified as
one.

Signed-off-by: Till Maas <opensource@till.name>