add markdownlint testenv (#119)
Add `markdownlint` testenv. This uses the github action image, similar
to how the `ansible-lint` testenv works.  This should make it easier
to add a markdownlint github action check.

NOTE: Still missing support for a markdownlint config file.  I'll note
that the network role has a customized .mdl that might not be suitable
for general purpose use.