use allowlist (#124)
* Use check-ansible-managed.sh script instead of inline bash

The primary problem is that `#` must be escaped with tox 4 but
not with tox 3.  It's too difficult to try to get consistent
quoting and escaping that works with both tox 3 and 4.  Instead,
move the shell command into a separate shell script file.

* Use allowlist instead of whitelist

tox 4 does not support whitelist, so must use allowlist
get rid of unused testenvs
Add support for ansible-core 2.15