plugin-scan - use explicit collection requirements
Install the collections from meta/collection-requirements.yml
and tests/collection-requirements.yml to override those found
in the `ansible` bundle.  We have to use an old ansible to
maintain compatability with jinja 2.7 and the collections
bundled there are not updated with newer versions of
ansible.posix, community.general, etc.

At some point when we drop compatability with EL7 and python2,
we can drop support for jinja 2.7 and use the latest ansible.

Signed-off-by: Rich Megginson <rmeggins@redhat.com>