ci: Drop support for Python 3.8
RHEL 8 has Python 3.6, RHEL 9 has Python ≥ 3.9, currently supported
Fedoras have ≥ 3.12. So none of the platforms we currently support have
3.8.

Dropping 3.8 is helpful because that is really hard to support wrt. the
`pkg_resources` → `importlib` migration. 3.8 doesn't yet have
`importlib.resources.files()`, which we need for that. Littering the
whole code with pylint and mypy overrides for this unknown API would be
error prone and ugly.