ci: Test Python 3.12 and 3.13 (#178)
* ci: Disable fast failing

This is annoying, as it serializes the errors of different python
versions.

* ci: Ignore https://data.safetycli.com/v/75180/f17/

We control which wheels we install (which is only setuptools), but we
don't control the pip version for the test, so ignore this. This
unbreaks the Python 3.6 test.

* ci: Drop support for Python 3.8

RHEL 8 has Python 3.6, RHEL 9 has Python ≥ 3.9, currently supported
Fedoras have ≥ 3.12. So none of the platforms we currently support have
3.8.

Dropping 3.8 is helpful because that is really hard to support wrt. the
`pkg_resources` → `importlib` migration. 3.8 doesn't yet have
`importlib.resources.files()`, which we need for that. Littering the
whole code with pylint and mypy overrides for this unknown API would be
error prone and ugly.

* fix: Prefer importlib.resources over pkg_resources

`pkg_resources` was deprecated in Python 3.12 and removed from 3.13 [1].
It still lives on in the `setuptools` package, but let's just move to
the current `importlib.resources` API, see the migration guide [2].

Still keep the `pkg_resources` code path as a fallback for Python 2.

[1] https://setuptools.pypa.io/en/latest/pkg_resources.html
[2] https://importlib-resources.readthedocs.io/en/latest/migration.html

* ci: Test Python 3.12 and 3.13

As the pkg_resources module is deprecated, the pylint env now needs an
explicit setuptools dependency to be able to check it.

Signed-off-by: Martin Pitt <mpitt@redhat.com>

* ci: HACK: Pin flake8 python to 3.11

With Python 3.12 and 3.13, flake8 fails:

```
   eps = importlib_metadata.entry_points().get(self.namespace, ())
          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
AttributeError: 'EntryPoints' object has no attribute 'get'
```

This is somehow related to https://github.com/PyCQA/flake8/issues/1701
and https://github.com/python/importlib_metadata/issues/406. We can't
use the former `import_metadata < 5` workaround any more, as flake8
imports this library from Python's stdlib (`importlib.metadata`), so the
external module gets ignored.

I don't know how to fix that. As the Python version that flake8 runs
with isn't very interesting, pin it to 3.11 for the time being.

---------

Signed-off-by: Martin Pitt <mpitt@redhat.com>