ci: Use package instead of file name for package resources (#181)
Unlike `pkg_resources`, `importlib.resources` cannot be called with a
module name inside of a package, like `tox_lsr.utils`, that fails with

> TypeError: 'tox_lsr.utils' is not a package

For some reason that doesn't break the tests here, but it does break in
consumers like the sudo role's tests.

So always request the package resources from the package intead of the
module.