ci: Move Python 2.7 test to Ubuntu 22.04 host, run python 3.6 test in UBI8 container (#182)
* ci: Move Python 2.7 test to Ubuntu 22.04 host

The "ubuntu-20.04" runner will be switched off in a week, and is
currently browning out.

See https://github.com/actions/runner-images/issues/11101.

* ci: Run python 3.6 test in UBI8 container

Python 3.6 is not available in newer Ubuntu releases any more, and
actions/setup-python isn't able to magically set this up either.

The only interesting platform for Python 3.6 is RHEL 8, so let's run
that test in an UBI RHEL 8 container.