container tests: Cleanups, fixes, and bootc mode (#188)
* runcontainer: Run playbooks with separate ansible commands

The test playbooks are meant to be independent and standalone. Running
all of them in the same `ansible-playbook` command leaks facts and other
state between them. Parallel mode already does that, this only affected
serial mode.

* runcontainer: Install python3-libdnf5 for Fedora ≥ 41

Otherwise the `package:` builtin fails with "Could not import the
libdnf5 python module".

* runcontainer: Outdent refresh_test_container()

Invert the "is the existing container current" check, so that the main
part of the function can be outdented by one level.

* runcontainer: Move setup.yml build into refresh_test_container()

That localizes the code and makes it easier to skip
refresh_test_container() for bootc scenarios.

* runcontainer: Use temporary workdir

This eases debugging, as all the files are in a single place with human
readable names. It also simplifies the cleanup traps.

* runcontainer: Split out run_podman()

This is an independent step and logic from running the playbooks. This
makes the functions more isolated, easier to read, and easier to replace
with a different implementation.

* runcontainer: Add bootc mode

If the image name contains "-bootc", run the test in bootc mode:

 * Skip the test container setup. We want our roles to work in the
   standard distribution bootc base images, admins won't have a "magic"
   base container either.

 * Use `buildah` instead of `podman`. This is what tools like
   https://github.com/ansible-community/ansible-bender do, and we don't
   want roles to make assumptions about a fully "running" container --
   everything that they do during container build should only change
   static files, not processes.

 * Consequently, adjust the plugin setup to install the buildah connector in
   bootc mode.

https://issues.redhat.com/browse/RHEL-85652