runcontainer: Disable ansible_become
When running firewalld container tests in GitHub actions, all the tests
mysteriously fail right at the start in `gather_facts` with:

> "ansible.legacy.setup": {"ansible_facts": {"failed": true, "module_stderr":
> "sudo: PAM account management error: Authentication service cannot retrieve authentication info\nsudo: a password is required

It's not really clear where this comes from -- neither the GitHub
running host nor the buildah (for bootc) / podman (for classic rpm)
container requires a sudo password.

However, we can entirely avoid that: We never expect (or want) to run
`sudo` in containers, as all playbooks are expected to run as root
already.