feat: Sort QEMU batch tests in ascending or descending order by image name
Commit 2369683 sorted the batch tests in ascending order by default to
make test results more predictable and reproducible locally.

However, this makes the tests prone to introducing dependencies/cleanup
failures. We want to run some (ideally: half of) the tests in reverse
order to uncover that but still keep the reproducibility.

Introduce a new batch order mode `imagehash` which maps the image name
to one bit and use ascending/descending mode according to it. This makes
e.g. "fedora-42" use ascending and "centos-10" use descending order.

For completeness, introduce a `descending` order as well, and rename the
existing `alphanum` to `ascending` (as `revalphanum` gets a bit
unwieldy).

https://issues.redhat.com/browse/RHELMISC-11615