runcontainer: Skip reboot tests
Running `reboot` in a podman system container does not work -- it just
shuts down and stops the container, but never restarts it. So skip these
tests during container runs.

Also skip them in container builds. This is a bit redundant as there is
also `tests::booted`, but it's easier and more symmetric.