runcontainer: Skip reboot tests; tests: fix mypy (#199)
* runcontainer: Skip reboot tests

Running `reboot` in a podman system container does not work -- it just
shuts down and stops the container, but never restarts it. So skip these
tests during container runs.

Also skip them in container builds. This is a bit redundant as there is
also `tests::booted`, but it's easier and more symmetric.

* tests: Move linter tox environment to setuptools >= 71

The most recent types-setuptools versions > 80.4 are meant for
setuptools >= 71 which include typing. So they don't work any more with
older setuptools versions.

Replace types-setuptools with a recent setuptools version.