doc: Add "Integration test setup" section, update OS names, fix test playbook location (#205)
* doc: Add "Integration test setup" section

Show all the commands to get going in a single section. This repeats the
`pip install` command, but it's not going to change often, and it's
worth having a single section to refer to.

* doc: Move example OS and env to ansible-core-2.17 CentOS 10

CentOS 10 ages slower than Fedora releases. Also harmonize the tox
environments with the OS to match our regular test matrix.

* doc: Fix test playbook location

That was missing the tests/ directory prefix in a couple of places.